+++ mirrord/intproxy/src/proxies/outgoing.rs
@@ -16,7 +16,7 @@ use crate::{
         ProxyToLayerMessage,
     },
     request_queue::{RequestQueue, RequestQueueEmpty},
-    ProxyMessage,
+    session::ProxyMessage,
 };
 
 mod interceptor;
@@ -111,6 +111,7 @@ impl OutgoingProxy {
     /// Passes the data to the correct [`Interceptor`] task.
     /// Fails when the agent sends an error, because this error cannot be traced back to an exact
     /// connection.
+    #[tracing::instrument(level = "trace", skip(self))]
     async fn handle_agent_read(
         &mut self,
         read: RemoteResult<DaemonRead>,
@@ -141,6 +142,7 @@ impl OutgoingProxy {
     /// Handles agent's response to a connection request.
     /// Prepares a local socket and registers a new [`Interceptor`] task for this connection.
     /// Replies to the layer's request.
+    #[tracing::instrument(level = "trace", skip(self, message_bus))]
     async fn handle_connect_response(
         &mut self,
         connect: RemoteResult<DaemonConnect>,
@@ -197,6 +199,7 @@ impl OutgoingProxy {
     }
 
     /// Saves the layer's request id and sends the connection request to the agent.
+    #[tracing::instrument(level = "trace", skip(self, message_bus))]
     async fn handle_connect_request(
         &mut self,
         message_id: MessageId,
